+++ moabb/datasets/utils.py
@@ -20,7 +20,8 @@ def dataset_search(paradigm, multi_session=False, events=None,
     criteria are:
 
     events: list of strings
-    total_classes: int, total number of classes, will either truncate or choose from events. Defaults to 100 to keep all classes
+    total_classes: int, total number of classes, will either truncate or choose
+    from events. Defaults to 100 to keep all classes.
     has_all_events: bool, skip datasets that don't have all events in events
     multi_session: bool, if True only returns datasets with more than one
     session per subject. If False return all
@@ -34,8 +35,8 @@ def dataset_search(paradigm, multi_session=False, events=None,
     n_classes = total_classes
     if events is not None and has_all_events:
         n_classes = len(events)
-    assert paradigm in ['imagery','p300']
-    if paradigm=='p300':
+    assert paradigm in ['imagery', 'p300']
+    if paradigm == 'p300':
         raise Exception('SORRY NOBDOYS GOTTEN AROUND TO THIS YET')
     for type_d in dataset_list:
         d = type_d()
@@ -49,8 +50,8 @@ def dataset_search(paradigm, multi_session=False, events=None,
             if events is None:
                 # randomly keep n_classes events
                 for k in d.event_id.keys():
-                   if len(keep_event_dict) < n_classes:
-                       keep_event_dict[k] = d.event_id[k]
+                    if len(keep_event_dict) < n_classes:
+                        keep_event_dict[k] = d.event_id[k]
             else:
                 n_events = 0
                 for e in events:
@@ -58,13 +59,13 @@ def dataset_search(paradigm, multi_session=False, events=None,
                         break
                     if e in d.event_id.keys():
                         keep_event_dict[e] = d.event_id[e]
-                        n_events+=1
+                        n_events += 1
                     else:
                         if has_all_events:
                             skip_dataset = True
                 # don't want to use datasets with one valid label
                 if n_events < 2:
-                    skip_dataset=True
+                    skip_dataset = True
             if keep_event_dict and not skip_dataset:
                 d.selected_events = keep_event_dict
                 if len(channels) > 0:
@@ -75,9 +76,11 @@ def dataset_search(paradigm, multi_session=False, events=None,
                     out_data.append(d)
     return out_data
 
+
 def find_intersecting_channels(datasets, verbose=False):
     '''
-    Given a list of dataset instances return a list of channels shared by all datasets.
+    Given a list of dataset instances return a list of channels shared by all
+    datasets.
     Skip datasets which have 0 overlap with the others
 
     returns: set of common channels, list of datasets with valid channels
@@ -105,5 +108,5 @@ def find_intersecting_channels(datasets, verbose=False):
             print('Dataset {:s} has no recognizable EEG channels'.format(type(d).__name__))
     for d in dset_chans:
         allchans.intersection_update(d)
-    allchans = [s.replace('Z','z') for s in allchans]
+    allchans = [s.replace('Z', 'z') for s in allchans]
     return allchans, d
