+++ moabb/viz/results.py
@@ -21,7 +21,7 @@ class Results:
         """
         class that will abstract result storage
         """
-        import moabb.utils as ut
+        import moabb.datasets.utils as ut
         from moabb.contexts.base import BaseParadigm
         from moabb.evaluations.base import BaseEvaluation
         assert issubclass(evaluation_class, BaseEvaluation)
